Implemented currency search functionality by name/symbol, added 2 themes, fixed structure

+++ src/TestCrypto/Infrastructure/Responses/CurrencyResponse.cs
@@ -69,7 +69,7 @@ public record CurrencyResponse
     public decimal? Atl { get; init; }
 
     [JsonProperty("atl_change_percentage")]
-    public decimal? AtlChangePercentage { get; init; }
+    public double? AtlChangePercentage { get; init; }
 
     [JsonProperty("atl_date")]
     public DateTime AtlDate { get; init; }
